+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/extensions/dynamodb.py
@@ -62,13 +62,42 @@ def _conv_val_to_len(value) -> Optional[int]:
     return len(value) if value is not None else None
 
 
+_db_statement_operations = {
+    "BatchGetItem",
+    "BatchWriteItem",
+    "DeleteItem",
+    "GetItem",
+    "PutItem",
+    "Query",
+    "Scan",
+    "UpdateItem",
+}
+
+_sanitized_keys = (
+    "Item",
+    "Key",
+    "KeyConditionExpression",
+    "ExpressionAttributeValues",
+    "FilterExpression",
+)
+_sanitized_value = {"?": "?"}
+
+
+# pylint: disable=C0103
+def _conv_params_to_sanitized_str(params) -> str:
+    p = params.copy()
+    for key in p:
+        if key in _sanitized_keys:
+            p[key] = _sanitized_value
+    return p
+
+
 ################################################################################
 # common request attributes
 ################################################################################
 
 _REQ_TABLE_NAME = ("TableName", _conv_val_to_single_attr_tuple)
 _REQ_REQITEMS_TABLE_NAMES = ("RequestItems", _conv_dict_to_key_tuple)
-
 
 _REQ_GLOBAL_SEC_INDEXES = ("GlobalSecondaryIndexes", _conv_list_to_json_list)
 _REQ_LOCAL_SEC_INDEXES = ("LocalSecondaryIndexes", _conv_list_to_json_list)
@@ -82,7 +111,6 @@ _REQ_ATTRS_TO_GET = ("AttributesToGet", None)
 _REQ_LIMIT = ("Limit", None)
 _REQ_SELECT = ("Select", None)
 _REQ_INDEX_NAME = ("IndexName", None)
-
 
 ################################################################################
 # common response attributes
@@ -350,11 +378,26 @@ class _DynamoDbExtension(_AwsSdkExtension):
     def __init__(self, call_context: _AwsSdkCallContext):
         super().__init__(call_context)
         self._op = _OPERATION_MAPPING.get(call_context.operation)
+        self._configuration = (
+            call_context.configuration.get_dynamodb_configuration()
+        )
 
     def extract_attributes(self, attributes: _AttributeMapT):
         attributes[SpanAttributes.DB_SYSTEM] = DbSystemValues.DYNAMODB.value
         attributes[SpanAttributes.DB_OPERATION] = self._call_context.operation
         attributes[SpanAttributes.NET_PEER_NAME] = self._get_peer_name()
+
+        if (
+            self._call_context.operation in _db_statement_operations
+            and self._call_context.params
+        ):
+            params = self._call_context.params
+            if self._configuration.get("sanitize_query"):
+                params = _conv_params_to_sanitized_str(params)
+
+            attributes[SpanAttributes.DB_STATEMENT] = _conv_dict_to_json_str(
+                params
+            )
 
         if self._op is None:
             return